requirements.txt
astroid==2.4.0
autopep8==1.5.2
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
entrypoints==0.3
flake8==3.7.9
future==0.18.2
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
line-bot-sdk==1.16.0
mccabe==0.6.1
numpy==1.18.3
pandas==1.0.3
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.5.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
selenium==3.141.0
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
